Move secure_scrub_memory to os_utils.cpp
